third_party/WebKit/LayoutTests/http/tests/htmlimports/import-dynamic-block-dynamic.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
</head>
<body>
<script>
(function() {
    var t = async_test('Dynamically added import blocks dynamic added imports.');

    function createLink(url, onload) {
        var link = document.createElement('link');
        link.rel = 'import';
        link.href = url;
        link.onload = onload;
        document.head.appendChild(link);
        return link;
    }

    var dynamicLink = createLink('resources/import-slow-child.cgi');
    var followingLink = createLink('resources/import-slow-child.cgi', function() {
        t.step(function() {
            assert_true(null != dynamicLink.import, 'Dynamic import'); 
            assert_true(null != followingLink.import, 'Following import'); 
            t.done();
        });
    });
})();
</script>
</body>
</html>
